Pequeños cambios
-Agregue un minimapa de la pista
-Le saque un box collider de mas q tenia el jeep
-Cambie como se muestra la velocidad

+++ ProyectoUnityVJ/Assets/Scripts/TestController.cs
@@ -81,6 +81,6 @@ public class TestController : MonoBehaviour
     protected void UIText()
     {
         //_localVelocity = transform.InverseTransformDirection(_rb.velocity);
-        speedText.text = "Speed: " + currentSpeed;
+        speedText.text = "Speed: " + (int)currentSpeed;
     }
 }
